Fixed value for reach snippet average score

+++ app/code/community/Yotpo/Yotpo/sql/yotpo_setup/mysql4-install-1.6.2.php
@@ -7,7 +7,7 @@ $installer->run("
     CREATE TABLE `yotpo_rich_snippets` (
       `rich_snippet_id` int(11) NOT NULL auto_increment,
       `product_id` int(11) NOT NULL,
-      `store_id` int(11) NOT NULL,
+      `store_id` float(11) NOT NULL,
       `average_score` int(11) NOT NULL,
       `reviews_count` int(11) NOT NULL,
       `expiration_time` timestamp NOT NULL default CURRENT_TIMESTAMP,
